удаляет моковые данные
