--- src/mock/points.js
@@ -1,33 +1,0 @@
-import { getRandomInteger } from '../helpers/utils';
-import { getRandomStartDate, getRandomEndDate } from '../helpers/dates';
-
-const Price = {
-  MIN: 1000,
-  MAX: 2000,
-};
-
-const defaultPoint = {
-  id: null,
-  basePrice: 0,
-  dateFrom: null,
-  dateTo: null,
-  destination: '',
-  isFavorite: false,
-  offers: [],
-  type: 'Flight',
-};
-
-function createPoint (destinationId, type, offerIds) {
-  return {
-    id: crypto.randomUUID(),
-    basePrice: getRandomInteger(Price.MIN, Price.MAX),
-    dateFrom: getRandomStartDate(),
-    dateTo: getRandomEndDate(),
-    destination: destinationId,
-    isFavorite: !!getRandomInteger(0, 1),
-    offers: offerIds,
-    type: type,
-  };
-}
-
-export { defaultPoint, createPoint };
